Use updateValue in PlanOutOpSimple
Previously, updating values using setter method via subscripting
will remove the key provided that the value is nil. Using the
updateValue method will preserve the nil value instead.

The nil value needs to be preserved, because operators need the
differentiation whether the nil value is due to (1) the key does
not exist, or (2) the key *does* exist but the value is null.

Most operators would throw if the key does not exist, while some
operators (such as Length) allows nil values.