Fix sampling procedure in DefaultSegmentAllocator
The DefaultSegmentAllocator uses Set-backed store, which do not
guarantee the order of the elements contained within the collection.

Since Sample operator depends on the order of the elements within
the collection, it is expected that given the same input and the
same order, it should always produce the same samples.

The fix is done by sorting the Set before it is supplied to the
sampler function.